backend/requirements-minimal.txt
# Core dependencies for UniGo Backend
fastapi>=0.110.0
uvicorn>=0.25.0
motor>=3.3.0
pymongo>=4.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
email-validator>=2.0.0
starlette>=0.37.0
